Use Underline navigation as part of the drafty layout
