+++ docs/content/components/test.mdx
@@ -1,17 +1,11 @@
 ---
 title: Button
 status: Alpha
-source: https://github.com/primer/doctocat/blob/master/theme/src/components/do-dont.js
 figma: https://github.com/primer/doctocat/blob/master/theme/src/components/do-dont.js
 ---
 
-import {useState} from 'react'
 import {DraftyLayout} from '@primer/gatsby-theme-doctocat'
 export default DraftyLayout
-import {UnderlineNav} from '@primer/react'
-import {SubNav} from '../../../theme/src/components/sub-nav'
-
-<SubNav />
 
 ## Usage
