+++ theme/gatsby-node.js
@@ -18,7 +18,7 @@ exports.createPages = async ({graphql, actions}, themeOptions) => {
         nodes {
           fileAbsolutePath
           rawBody
-          tableOfContents(maxDepth: 4)
+          tableOfContents(maxDepth: 3)
           parent {
             ... on File {
               relativeDirectory
